requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai>=1.40.0
python-docx